Adicionados testes com números negativos e corrigido problema de precisão da cauculadora.

+++ JavaApplication/src/javaapplication/Calculadora.java
@@ -8,28 +8,28 @@ package javaapplication;
 public class Calculadora {
 
     // atributo
-    private int resultado = 0;
+    private double resultado = 0;
 
     // método somar
-    public double somar(int n1, int n2) {
+    public double somar(double n1, double n2) {
         resultado = n1 + n2;
         return resultado;
     }
 
     // método subtrair
-    public double subtrair(int n1, int n2) {
+    public double subtrair(double n1, double n2) {
         resultado = n1 - n2;
         return resultado;
     }
 
     // método multiplicar
-    public double multiplicar(int n1, int n2) {
+    public double multiplicar(double n1, double n2) {
         resultado = n1 * n2;
         return resultado;
     }
 
     // método dividir
-    public double dividir(int n1, int n2) {
+    public double dividir(double n1, double n2) {
         resultado = n1 / n2;
         return resultado;
     }
